Cython/requirements.txt
flask==2.0.1
requests==2.31.0
Pillow==10.0.0 